removed unused services, repositories, and refactored UserInterestService to streamline logic

+++ src/main/java/dev/kuku/interestcalculator/models/entities/UserInterestEntity.java
@@ -5,5 +5,5 @@ import org.yaml.snakeyaml.util.Tuple;
 import java.util.Map;
 
 ///  userId -> topics & score, timestamp
-public record UserInterestEntity(String userId, Map<String, Tuple<Integer, Long>> topics) {
+public record UserInterestEntity(String userId, Map<String, Tuple<Long, Long>> topics) {
 }
